// src/generated/rustCore/SceneData.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CameraState } from "./CameraState";
import type { LayerState } from "./LayerState";
import type { StageCamera3D } from "./StageCamera3D";
import type { TimelineObject } from "./TimelineObject";

/**
 * `src/types.ts` の `SceneData` と同形。
 */
export type SceneData = { id: string, name: string, duration: number, layers: Array<LayerState>, objects: Array<TimelineObject>, camera: CameraState, stageCamera3D: StageCamera3D, };